+++ src/index.js
@@ -3,7 +3,7 @@ import AdView from "./AppLovinMAXAdView";
 
 const { AppLovinMAX } = NativeModules;
 
-const VERSION = "2.0.1";
+const VERSION = "2.0.2";
 
 /**
  * This enum represents whether or not the consent dialog should be shown for this user.