+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Versions
 
-## x.x.x
+## 2.0.2
     * Remove `getAdInfo(adUnitId)` API in lieu of ad callbacks.
     * Return more data in ad callbacks in addition to `ad.adUnitId` (e.g. `adInfo.creativeId`, `adInfo.networkName`, `adInfo.placement`, `adInfo.revenue`).
 ## 2.0.1